Challenges/Intigriti-CTF/Web/bounty_repo/admin/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2